requirements.txt
openai==1.75.0
python-dotenv == 1.0.1
Flask == 3.1.0
tika==3.1.0
flask-cors